requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
dateparser==1.2.0
Django==5.1.1
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-elasticsearch-dsl==8.0
django-timezone-field==7.0
elastic-transport==8.15.1
elasticsearch==8.15.1
elasticsearch-dsl==8.15.4
fuzzywuzzy==0.18.0
imaplib2==3.6
kombu==5.4.2
Levenshtein==0.26.0
packaging==24.1
pdf2image==1.17.0
pillow==10.4.0
prompt_toolkit==3.0.48
pytesseract==0.3.13
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.26.0
pytz==2024.2
RapidFuzz==3.10.0
redis==5.1.0
regex==2024.9.11
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
